requirements.txt
Django
djangorestframework
python-decouple
django-cors-headers
drf-yasg
setuptools
django-filter
pymysql
whitenoise
